t/lib-gitweb: test against the build version of gitweb
When testing gitweb we set up the CGI script as "gitweb.perl", which is
the source file of the build target "gitweb.cgi". This works alright as
long as we run in-tree tests. But we're about to make out-of-tree tests
a reality, and there things will break because "gitweb.perl" will not be
found in the build directory.

Fix this by using "gitweb.cgi" instead.

Signed-off-by: Patrick Steinhardt <ps@pks.im>
Signed-off-by: Junio C Hamano <gitster@pobox.com>