pack-bitmap.c: do not pass `pack_pos` to `try_partial_reuse()`
The function `try_partial_reuse()` uses its `pack_pos` parameter to
determine if:

  - the object specified by that position is in bounds of the pack
    it's operating on, and

  - in the case that the specified object is an OFS_DELTA (and we're
    operating on a single-pack reachability bitmap) to ensure that the
    position of the base object occurs earlier than the delta

But neither of these checks are necessary. The bounds-check is
redundant because we are operating on bit positions which we know
correspond to objects in some pack, and the caller in
reuse_partial_packfile_from_bitmap_1() is smart enough to know to stop
processing bits when it is at the end of some pack.

Likewise, the delta dependency check is also unnecessary, because we
can use the object offsets within a single pack as a proxy for bit
positions in that pack's bitmap.

So let's avoid passing in this redundant piece of information, and
save one call to offset_to_pack_pos(), which is O(log N) in the number
of objects in the pack we're currently processing.

(This all comes from a discussion on a semi-related series from [1]).

[1]: https://lore.kernel.org/git/Zti0xrzCtpWScPjz@nand.local/

Suggested-by: Junio C Hamano <gitster@pobox.com>
Signed-off-by: Taylor Blau <me@ttaylorr.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>