pack-bitmap.c: compare `base_offset` to `delta_obj_offset`
When comparing the offsets of either half of a delta/base-pair, we
compare `base_offset` to `offset`.

There is nothing functionally wrong with that comparison, but it is
slightly confusing, since `offset` points to just after the delta
object's type header in the pack, whereas `base_offset` points to the
beginning of the header.

In practice, that distinction does not matter, and it is perfectly
fine to compare base_offset to offset.

But we already make a copy of `offset` before it is moved forward by
calling the function `unpack_object_header()`. So let's use that copy
(which points at the beginning of the delta object's header) in the
comparison so that we are comparing like quantities.

Signed-off-by: Taylor Blau <me@ttaylorr.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>